do some CPU testing with Vanity API

+++ Timer.cpp
@@ -137,7 +137,7 @@ uint32_t Timer::getSeed32() {
   return ::strtoul(getSeed(4).c_str(),NULL,16);
 }
 
-std::string Timer::getResult(char *unit, int nbTry, double t0, double t1) {
+std::string Timer::getResult(char *unit, size_t nbTry, double t0, double t1) {
 
   char tmp[256];
   int pIdx = 0;
@@ -151,7 +151,7 @@ std::string Timer::getResult(char *unit, int nbTry, double t0, double t1) {
 
 }
 
-void Timer::printResult(char *unit, int nbTry, double t0, double t1) {
+void Timer::printResult(char *unit, size_t nbTry, double t0, double t1) {
 
   printf("%s\n", getResult(unit, nbTry, t0, t1).c_str());
